requirements.txt
Flask>=2.3,<3.1
Flask-SQLAlchemy>=3.1,<3.2
SQLAlchemy>=2.0,<2.1
Pillow>=10.0
numpy>=1.26.4,<2.0
pandas>=2.2,<2.3
openpyxl>=3.1
matplotlib>=3.8,<3.10
scikit-learn>=1.3,<1.6
scikit-optimize>=0.9.0
scipy>=1.11,<1.14
requests>=2.31,<3.0
openai>=1.40.0,<2.0
httpx>=0.27,<0.28
python-dotenv>=1.0,<2.0
